crypto-analyzer/requirements.txt
requests==2.31.0
pandas==2.0.3
rich==13.5.2
python-dotenv==1.0.0